доработана страница с планом выплат

+++ src/main/java/ru/bendricks/piris/service/UserService.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import ru.bendricks.piris.model.ObligationType;
+import ru.bendricks.piris.model.RecordStatus;
 import ru.bendricks.piris.model.User;
 import ru.bendricks.piris.model.UserRole;
 import ru.bendricks.piris.repository.UserRepository;
@@ -17,6 +19,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final ObligationService obligationService;
 
     @PreAuthorize("hasRole('ADMIN')")
     public List<User> getAllUsers() {
@@ -29,7 +32,16 @@ public class UserService {
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    public void deleteUser(long id) {
+    public void deleteUser(long id) throws Exception {
+        var activeCredits = obligationService.getObligationsByUserId(id).stream()
+                .filter(obligation ->
+                        (obligation.getObligationType().equals(ObligationType.CREDIT)
+                                || obligation.getObligationType().equals(ObligationType.CREDIT_ANUIT))
+                                && (obligation.getStatus().equals(RecordStatus.ACTIVE)
+                                || obligation.getStatus().equals(RecordStatus.END_OF_SERVICE)))
+                .toList();
+        if (!activeCredits.isEmpty())
+            throw new Exception("Невозможно удалить клиента так как у него есть незакртые кредиты");
         userRepository.deleteById(id);
     }
 
@@ -42,7 +54,7 @@ public class UserService {
 
     @Transactional
     @PreAuthorize("hasRole('ADMIN')")
-    public User updateUser(User user){
+    public User updateUser(User user) {
         user.setPasswordHash(userRepository.findById(user.getId()).map(User::getPasswordHash).orElse(null));
         return userRepository.save(user);
     }
@@ -53,17 +65,17 @@ public class UserService {
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    public boolean isPassportIdUsed(String passportId){
+    public boolean isPassportIdUsed(String passportId) {
         return userRepository.findUserByPassportId(passportId).isPresent();
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    public boolean isPassportSerialUsed(String passportSerial){
+    public boolean isPassportSerialUsed(String passportSerial) {
         return userRepository.findUserByPassportSerial(passportSerial).isPresent();
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    public boolean isMobilePhoneNumberUsed(String mobilePhoneNumber){
+    public boolean isMobilePhoneNumberUsed(String mobilePhoneNumber) {
         return userRepository.findUserByMobilePhoneNumber(mobilePhoneNumber).isPresent();
     }
